Removal of Lombok
Lombok implementation removal.

I have also gone through and replaced things with inline methods and variables, lambdas, and simplified loops down, removed unnecessary guard clauses, and overall cleaned up every single class. This took a long time, please do remember to follow proper naming conventions, don't include unnecessary guard clauses, follow exception rules and comment rules, and please PLEASE remember to use the DIAMOND OPERATOR rather than just inferring RAW TYPES!!!

Thank you!!

+++ src/main/java/me/totalfreedom/totalfreedommod/AutoKick.java
@@ -46,9 +46,10 @@ public class AutoKick extends FreedomService
 
     private void autoKickCheck()
     {
+        // No type cast was provided, one has been supplied.
         final boolean doAwayKickCheck
                 = plugin.esb.isEnabled()
-                && ((server.getOnlinePlayers().size() / server.getMaxPlayers()) > autoKickThreshold);
+                && (((float)server.getOnlinePlayers().size() / (float)server.getMaxPlayers()) > autoKickThreshold);
 
         if (!doAwayKickCheck)
         {
